refactor: open pylint in pre-commit (#48)
