+++ envs/safety-gymnasium/safety_gymnasium/bases/__init__.py
@@ -14,5 +14,5 @@
 # ==============================================================================
 """Base classes."""
 
-from safety_gymnasium.bases.base_mujoco_task import BaseMujocoTask
-from safety_gymnasium.bases.base_task import BaseTask
+# from safety_gymnasium.bases.base_mujoco_task import BaseMujocoTask
+# from safety_gymnasium.bases.base_task import BaseTask
